requirements.txt
discord.py==1.7.3
nltk
bs4
lxml
cchardet